refactor(server): extract check_dependencies to utils/dependency_check.py
Task G of docs/current/server-py-module-split.md. Pure extraction,
verbatim move verified byte-identical against git history
(check_dependencies, 62 lines). Removed now-dead pathlib.Path import
from server.py (SEMANTIC_SEARCH_AVAILABLE and DEFAULT_DB_PATH stay --
still used by lifespan's DB-location log and main()'s feature-status
log). Per the spec's G-before-H sequencing note, server.py's import
is updated now so Task H's extraction of main() carries the already-
correct `from .utils.dependency_check import check_dependencies`
call site into the new cli.py.

Functional check runs the real extracted function against the live
environment (not mocked): confirms it correctly reports installed
dependencies, semantic search, scheduler availability, and database
directory writability, returning True. 7 targeted tests in
test_cli_entrypoint.py pass unchanged -- no test monkeypatches this
function's globals, so no repointing was needed here.
